refactor: 중복 되는 코드 mapper 메서드로 빼줍니다
